new:  better support for custom mime-types
Do not read the body or try to decode the data when content-type or accept header is set to be one of the registered custom mime type.

This means the Data property of elemental.Request will not be populated. The handler must deal with reading the body from the internal http.Request.