fix regression error with map equality tests in push_config
Test code uses smarty Convey to check for two maps not equal, but
internally smarty uses reflect.DeepEqual which returns true, but the
test code expects it to return false, even though the map[] contents are
the same (deeply equal) but the maps are uniquely different in memory.
Maybe in the past this was returned as not equal, but today they return
as equal.  So I changed it to use reflect.ValueOf to validate the actual
objects are not the same.

it fixes the following error
* /home/runner/work/elemental/elemental/push_config_test.go
Line 508:
Expected     'map[i1:[create delete] i2:[create delete]]'
to NOT equal 'map[i1:[create delete] i2:[create delete]]'
(but it did)!