Updates (#139)
* Remove cov

* Update build-go.yaml

* Pin action versions

* Try using golangci-lint v1.64.8 (pre 2.0)

* Try using golangci-lint v1.54.2 (earlier, pre 2.0), and update deps

* Configure linters, particularly revive in golangci-lint v2 for Go 1.24

* Linting

* Linter configs

* No really v2

* Show version

* Linting

* Try disabling a few more cases

* Linting

* Linting

* run with linter for unchecked-type-assertion enabled

* fix regression error with map equality tests in push_config

Test code uses smarty Convey to check for two maps not equal, but
internally smarty uses reflect.DeepEqual which returns true, but the
test code expects it to return false, even though the map[] contents are
the same (deeply equal) but the maps are uniquely different in memory.
Maybe in the past this was returned as not equal, but today they return
as equal.  So I changed it to use reflect.ValueOf to validate the actual
objects are not the same.

it fixes the following error
* /home/runner/work/elemental/elemental/push_config_test.go
Line 508:
Expected     'map[i1:[create delete] i2:[create delete]]'
to NOT equal 'map[i1:[create delete] i2:[create delete]]'
(but it did)!

---------

Co-authored-by: james-otten-pan <69821911+james-otten-pan@users.noreply.github.com>
Co-authored-by: Brian Onn <bonn@paloaltonetworks.com>